server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.2.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzlocal==5.3.1
uritemplate==4.1.1
yarl==1.19.0
